Fixed lesson logic to use new lesson items
